maker: mod.rs 测试归位 + 拆掉隐藏 prelude（#261）
mod.rs 原本是其它模块的测试垃圾场（~670 行测的是 recovery/notify/model/
runtime 的内部），并因此兼任隐藏 prelude——头部若干导入自己不用，只为
`use super::*` 的下游服务，还得靠 `#[cfg(test)] use` 再导入块补齐。

归位：
- MakerExit 各条退出文案 + 仓位符号归一 + is_maker_order → model.rs
- validate_reconnect_snapshot 三个用例 → recovery.rs
- validate_alert_thresholds → startup.rs（新建 test 模块）
- webhook_body_shapes → notify.rs
- 三个 order-response 用例 → runtime/tests/order_events.rs
- 三个 mockito 端到端用例 → recovery.rs::live_gate_tests（它们驱动的都是
  recovery 的入口；controlled-disconnect 跨两半，注释说明了为什么它要
  额外导入 runtime 的 apply_order_responses）

跨 crate 重复：5 个用例重复验证 standx-maker 的纯函数。但直接删会丢覆盖
——CLI 版本带了 maker 版本没有的用例（qty 为 0/极小余量/浮点余量、
恰好等于上限的空头、pending_covers_slot 的错边错档）。先把这些补进
ownership.rs（顺手拆成三个按语义命名的测试），再删 CLI 侧重复。
PositionAlertAnchor 的两个 CLI 用例严格弱于 risk.rs 已有的四个，直接删。

隐藏 prelude 拆除：删掉 3 个 `#[cfg(test)] use` 再导入
（apply_order_responses / new_maker_rest_client / webhook_body）和
mod.rs 里 5 个只为测试存在的导入；runtime/tests 与 runtime_flow.rs 改为
显式导入 OrderSide / new_maker_rest_client。

mod.rs 1031 → 144 行，不再含任何测试或 cfg(test) 导入。
CLI 211 + maker 189 全绿，fmt --check / clippy -D warnings 干净。

注：integration_tests 里 market symbols 的几个用例打真实行情接口，
并行跑会偶发失败（--test-threads=1 全绿），与本次改动无关。
