+++ crates/standx-cli/src/commands/maker/startup.rs
@@ -522,3 +522,23 @@ pub(super) async fn run_startup(
         live_session,
     })
 }
+
+#[cfg(test)]
+mod tests {
+    use super::validate_alert_thresholds;
+
+    #[test]
+    fn alert_thresholds_reject_silent_disable_and_unfireable_ranges() {
+        // Baseline: all valid.
+        assert!(validate_alert_thresholds(50.0, 80.0, 20.0, 3600.0).is_ok());
+        // Zero everywhere means "disabled" and is allowed.
+        assert!(validate_alert_thresholds(0.0, 0.0, 0.0, 0.0).is_ok());
+        // Negative thresholds silently disable the guard.
+        assert!(validate_alert_thresholds(-1.0, 80.0, 20.0, 3600.0).is_err());
+        assert!(validate_alert_thresholds(50.0, -1.0, 20.0, 3600.0).is_err());
+        assert!(validate_alert_thresholds(50.0, 80.0, 20.0, -1.0).is_err());
+        // Percentages above 100 can never fire.
+        assert!(validate_alert_thresholds(50.0, 170.0, 20.0, 3600.0).is_err());
+        assert!(validate_alert_thresholds(50.0, 80.0, 170.0, 3600.0).is_err());
+    }
+}
